GlobalEnums and update to RS232
