+++ vvvv45/src/nodes/plugins/Devices/RS232Node.cs
@@ -11,30 +11,6 @@ using VVVV.Core.Logging;
 
 namespace VVVV.Nodes.Devices
 {
-	[Startable]
-	public class Rs232PortManager : IStartable
-	{
-		public const string COM_PORT_ENUM_NAME = "Rs232Node.ComPort";
-		
-        public static void UpdatePortList()
-        {
-            var portNames = SerialPort.GetPortNames()
-                .Where(n => n.StartsWith("com", StringComparison.InvariantCultureIgnoreCase))
-                .ToArray();
-            EnumManager.UpdateEnum(COM_PORT_ENUM_NAME, portNames.Length > 0 ? portNames[0] : string.Empty, portNames);
-        }
-		
-		#region IStartable implementation
-		public void Start()
-		{
-			UpdatePortList();
-		}
-		public void Shutdown()
-		{
-		}
-		#endregion
-	}
-	
     [PluginInfo(Name = "RS232", Category = "Devices", AutoEvaluate = true)]
     public class Rs232Node : IDisposable, IPluginEvaluate, IPartImportsSatisfiedNotification
     {
@@ -60,7 +36,7 @@ namespace VVVV.Nodes.Devices
         public ISpread<bool> BreakStateIn;
         [Input("Update Port List", IsSingle=true, IsBang=true, Visibility = PinVisibility.Hidden)]
         public IDiffSpread<bool> UpdatePortListIn;
-        [Input("Port Name", EnumName = Rs232PortManager.COM_PORT_ENUM_NAME)]
+        [Input("Port Name", EnumName = GlobalEnumManager.COM_PORT_ENUM_NAME)]
         public ISpread<EnumEntry> ComPortIn;
         [Input("Enabled")]
         public ISpread<bool> EnabledIn;
@@ -105,7 +81,7 @@ namespace VVVV.Nodes.Devices
             RiStateOut.SliceCount = spreadMax;
             BreakStateOut.SliceCount = spreadMax;
 
-            if (UpdatePortListIn.IsChanged && UpdatePortListIn[0]) Rs232PortManager.UpdatePortList();
+            if (UpdatePortListIn.IsChanged && UpdatePortListIn[0]) GlobalEnumManager.UpdatePortList();
 
             FPorts.Resize(spreadMax, CreatePort, DestroyPort);
             DataOut.ResizeAndDispose(spreadMax, () => new MemoryStream());
