+++ vvvv45/src/nodes/plugins/NativePlugins.csproj
@@ -86,6 +86,7 @@
     <Compile Include="Animation\ParticlesThreaded\ParticleSystemNode.cs" />
     <Compile Include="AssemblyInfo.cs" />
     <Compile Include="Devices\RS232Node.cs" />
+    <Compile Include="GlobalEnums.cs" />
     <Compile Include="Layer\DrawText\DrawText.cs" />
     <Compile Include="Layer\DrawText\DrawTextLegacy.cs" />
     <Compile Include="Network\Clock.cs" />
