+++ vvvv45/src/vvvv45.sln
@@ -1,7 +1,9 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-# SharpDevelop 4.4
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2012
+# SharpDevelop 5.0
+VisualStudioVersion = 12.0.20827.3
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "core", "core", "{704B6D46-8445-4315-8128-F331143A113E}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Utils3rdParty", "core\Utils3rdParty\Utils3rdParty.csproj", "{87863121-C124-4585-B3BB-7706F03ECC64}"
@@ -255,16 +257,16 @@ Global
 		{98936436-EA28-437E-A2A4-CC7CAA532658} = {7FE983D0-4392-451C-9BAE-2CA9036FAB04}
 		{304E1C6A-585F-4748-8E3B-33558E35B0D4} = {7FE983D0-4392-451C-9BAE-2CA9036FAB04}
 		{90D41549-2657-4E4A-AF4E-3214A007C991} = {C11D1E7D-56CC-4DE9-A43A-308B39408D50}
-		{992F311E-FA5D-4732-88EA-7296524FA1F4} = {C11D1E7D-56CC-4DE9-A43A-308B39408D50}
 		{A2B1BF86-17C2-4122-AB0A-1789B19793E5} = {90D41549-2657-4E4A-AF4E-3214A007C991}
+		{8B9D2C28-4394-4C86-A14C-B0021EC29089} = {A2B1BF86-17C2-4122-AB0A-1789B19793E5}
 		{716F4773-5590-4DD6-A9BC-3FE905D22DA5} = {90D41549-2657-4E4A-AF4E-3214A007C991}
 		{763EEECF-9A9C-4FF8-AB6D-2854E7A510D9} = {90D41549-2657-4E4A-AF4E-3214A007C991}
-		{7C320C90-D64B-4B27-9234-A988141E6775} = {90D41549-2657-4E4A-AF4E-3214A007C991}
-		{E7B99BB8-5720-45FA-A395-0F9FC8212117} = {90D41549-2657-4E4A-AF4E-3214A007C991}
-		{8B9D2C28-4394-4C86-A14C-B0021EC29089} = {A2B1BF86-17C2-4122-AB0A-1789B19793E5}
 		{639A1C44-DA07-4F08-8A90-8B8845515BF4} = {763EEECF-9A9C-4FF8-AB6D-2854E7A510D9}
+		{7C320C90-D64B-4B27-9234-A988141E6775} = {90D41549-2657-4E4A-AF4E-3214A007C991}
 		{E0C5B55A-37EA-48D6-95B9-5EEF07C4AD8E} = {7C320C90-D64B-4B27-9234-A988141E6775}
+		{E7B99BB8-5720-45FA-A395-0F9FC8212117} = {90D41549-2657-4E4A-AF4E-3214A007C991}
 		{E051BB88-5211-4EF7-962B-59657D171297} = {E7B99BB8-5720-45FA-A395-0F9FC8212117}
+		{992F311E-FA5D-4732-88EA-7296524FA1F4} = {C11D1E7D-56CC-4DE9-A43A-308B39408D50}
 		{A63D2EA4-D87C-459B-B57F-19A76A7ADAF0} = {992F311E-FA5D-4732-88EA-7296524FA1F4}
 	EndGlobalSection
 EndGlobal
